pyproject.toml
[tool.poetry]
name = "cg"
version = "0.1.0"
description = ""
authors = ["scc <scc@teamt5.org>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pylint = "^2.17.4"
autopep8 = "^2.0.2"
argparse = "^1.4.0"
openai = "^0.27.7"
gitpython = "^3.1.31"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
